Rename resources to v2
