+++ service/resource/namespacev2/desired_test.go
@@ -1,4 +1,4 @@
-package namespacev1
+package namespacev2
 
 import (
 	"context"
