+++ service/resource/namespacev2/resource.go
@@ -1,4 +1,4 @@
-package namespacev1
+package namespacev2
 
 import (
 	"github.com/giantswarm/microerror"
@@ -10,7 +10,7 @@ import (
 
 const (
 	// Name is the identifier of the resource.
-	Name = "namespacev1"
+	Name = "namespacev2"
 )
 
 // Config represents the configuration used to create a new deployment resource.
