+++ service/resource/s3bucketv2/desired_test.go
@@ -1,4 +1,4 @@
-package s3bucketv1
+package s3bucketv2
 
 import (
 	"context"
