+++ service/resource/s3bucketv2/spec.go
@@ -1,4 +1,4 @@
-package s3bucketv1
+package s3bucketv2
 
 import (
 	"github.com/aws/aws-sdk-go/service/s3"
