+++ service/resource/s3objectv2/create_test.go
@@ -1,4 +1,4 @@
-package s3objectv1
+package s3objectv2
 
 import (
 	"bufio"
