+++ service/resource/s3objectv2/spec.go
@@ -1,4 +1,4 @@
-package s3objectv1
+package s3objectv2
 
 import (
 	"github.com/aws/aws-sdk-go/service/kms"
